requirements.txt
# Base requirements file
Django==3.2.6
django-environ==0.4.5
gunicorn==20.1.0
Collectfast==2.2.0
django-settings-export==1.2.1
psycopg2==2.9.1
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
django-cors-headers
django-model-utils
djangorestframework-datatables
drf-nested-routers